requirements.txt
Pillow
mysql-connector-python
pandas
numpy==1.19.3